requirements copy.txt
# Core
django==4.2.9
django-allauth==0.60.1
PyJWT==2.8.0
python-dotenv==1.0.1
str2bool==1.1
django-quill-editor==0.1.40

# Misc
celery==5.3.4
redis==5.0.1
hiredis==2.2.3
pillow==10.2.0

# Tools
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
django-debug-toolbar==4.2.0
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0
cryptography==42.0.2

# DB Layer
psycopg2-binary
# mysqlclient

# ENTERPRISE Version
#  - CVS to Model
#  - Loader  
pandas==2.2.0
google-auth==2.26.2
google_auth_oauthlib==1.2.0
google-api-python-client==2.147.0

openai==1.13.3
youtube-transcript-api
str2bool
tiktoken
trafilatura
pydantic
exa_py
geopy
langchain
langchain_community
langchain_core
langchain_openai==0.0.2.post1
Requests==2.32.3
python-dotenv
beautifulsoup4==4.12.2
crewai==0.65.2
crewai_tools
embedchain
langsmith==0.0.87
